FEAT: finalize how DoSeparateHPI is specified
Last commit on this issue.

Relevant files have been modified to allow for the DoSeparateHPI option to only need setting once.

A logical flag, in either the NOAA or SME (if doUseAEforHP) file definition of UAM.in, tells GITM whether to read HPI's into hpi_ or both hpi_ and hpi_sh_

Aurora.Earth.f90 then decides whether to scale aurora separately by hemisphere or not.

UAM.in templates & inputs' documentation are updated to reflect this.