STY:  Begin refactoring src/aurora.Earth.f90
This commit modifies arc/aurora.Earth.f90 and src/ModSources.f90 to
accomadate refactoring of Fang calculations.

If Fang is used for energy deposition, it is called in 2 steps. First, it
is initialized (initialize_fang). This meant ModSources had to be changed
in addition to aurora.

Then it is run in calc_fang_rates. This is in a loop over all grid cells,
so returned is the energy deposition in a single lon/lat over all altitudes.

Tested running with & without Fang, with & without HPI input and I couldn't break anything.