[BUG]: do not use time info when reading indices in the SWMF
GITM is initialized without start/end times when running in component mode. So we cannot check the validity of indices files.

able to be read:
- MHD Indices
- HPI
- F107

Currently, this means that SME and OMNI AP cannot be read, since their readers need input times.

THIS IS STILL A BUG, but at least we can run without error.