BUG: Turn off automatic test runs, for now.
GITM doesn't work with MPICH, apparently. Same error on my laptop with mpich. Switched to openmpi, no error!

This is not related to this pull request.... Looking back through the test logs, it has not been running successfully for a while. But not the script is smart enough to recognize when it is supposed to fail.

I will debug this later. Easy solution is to switch to openmpi, but we need to get mpich working at some point. So I am pushing the problem down the road